Renamed SignUpPage and SignUpForm components to index.js. Restructured scss files imports
