+++ src/pages/SignUpPage/index.js
@@ -4,11 +4,9 @@ import { FormattedMessage } from 'react-intl';
 
 import { useSession, useDispatch } from 'hooks';
 import { signUp } from 'state/actions/userActions';
-import SignUpForm from 'components/user/SignUpForm/SignUpForm';
-import SocialMediaContainer from 'components/common/SocialMediaContainer/SocialMediaContainer';
+import SignUpForm from 'components/user/SignUpForm';
+import SocialMediaContainer from 'components/common/SocialMediaContainer';
 import routes from 'constants/routesPaths';
-
-import './sign-up-page.scss';
 
 const SignUpPage = () => {
   const { authenticated } = useSession();
