+++ src/styles/styles.scss
@@ -3,8 +3,10 @@
 
 @import '../components/common/SocialMediaContainer/social-media-container.scss';
 @import '../components/user/LoginForm/login-form.scss';
+@import '../components/user/SignUpForm/sign-up-form.scss';
 
 @import '../pages/LoginPage/login-page.scss';
+@import '../pages/SignUpPage/sign-up-page.scss';
 
 html,
 body {
